requirement.txt
ostorlab[agent]
rich
fastmcp
pydantic_ai[google]
